fix: warn if not non object given to object option
Warn and ignore
